Detect POD sections like perl would
perl considers that a POD section begins with a line that starts by "=X",
X being any alphabetic character, and ends by "=cut" followed by a
non-alphabetic character. This commit makes Module::Metadata follow this
strategy.

Note that there's still a gotcha : "=X" must be a statement to start a POD
section. This means that in

    my $x
    =pod
    ;

'pod' is treated as a bareword, hence is compiled to either a string
literal or a function call depending on whether a pod() sub has been
previously declared in the current package. Module::Metadata cannot do
anything about this situation.

This commit partially reverts the previous commit a4aafbc2.