Remove warning for when modules don't do $VERSION = eval $VERSION.
In a non-trivial number of cases this will be triggered on code one
doesn't have any control over, and thus can't really fix--the verbosity
therefore doesn't actually provide any benefit to the user. In the
original use case (Module::Build::ModuleInfo) this sort of warning makes
good sense as a sanity check for authors. It has no place in a general
purpose library, though.

... not to mention the bloody thing was an unless/else block. Seriously?
And the warning did not at all communicate what the actual problem was.
Bad UX all around.