dzilify this distribution!
Some very small zero-impact modifications to tests have been made in the
interests of keeping dzil plugins happy enough to parse all this content.

e.g. a comment is altered to avoid appearing to assign to $VERSION (see also
rjbs/Dist-Zilla#322). I don't think this is a valid bit of code anyway - if
testing a lcoal repo, lib or blib should already be in @INC, and if it isnt,
we really ought to fail.  @dagolden added this in commit
Perl-Toolchain-Gang/Module-Build@801be221, so clarification is needed.