Add $self->{inc} to @INC whilst evaluating version
WWW::Scripter defines multiple packages in the same file, some of those
packages contain a block of text that Module::Metadata tries to evaluate
to get a version. The line that Module::Metadata tries to evaluate is
"use WWW::Scripter; use WWW::Scripter; $VERSION = $WWW'Scripter'VERSION;"
When using Carton to install WWW::Scripter this causes Module::Metadata
to error with "Can't find module". If we add $self->{inc} to @INC before
evaluating the version line then Perl looks in the right places for the
WWW::Scripter module.